+++ src/Debug.php
@@ -2,6 +2,9 @@
 
 namespace Debugger;
 
+use Debugger\Tools\Js;
+use Debugger\Tools\Cmd;
+use Debugger\Tools\Log;
 use Debugger\Tools\Lorium;
 use Debugger\Tools\Timing;
 use Debugger\Tools\DisplayData;
@@ -49,6 +52,39 @@ class Debug {
 
     public static object $lorium;
 
+    /**
+     * Object for using a number of terminal tools.
+     * 
+     * @var object  $cmd
+     * 
+     * @access  public
+     * @since   1.0.1
+     */
+
+    public static object $cmd;
+
+    /**
+     * Object for using a number of javascript tools.
+     * 
+     * @var object  $js
+     * 
+     * @access  public
+     * @since   1.0.1
+     */
+
+    public static object $js;
+
+    /**
+     * Object for using a number of logging tools.
+     * 
+     * @var object  $log
+     * 
+     * @access  public
+     * @since   1.0.1
+     */
+
+    public static object $log;
+
 
     /**
      * Constructor method, should be placed in the autoloader or called 
@@ -59,9 +95,12 @@ class Debug {
      */
     
     public static function __constructStatic() {
-        self::$timer = new Timing;
+        self::$timer   = new Timing;
         self::$display = new DisplayData;
-        self::$lorium = new Lorium;
+        self::$lorium  = new Lorium;
+        self::$cmd     = new Cmd;
+        self::$js      = new Js;
+        self::$log     = new Log;
     }
 
 }