+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # PHP DEBUGGIN TOOLS - CHANGELOG
 
+## Version 1.0.2 - 2022-04-27
+
+### Added
+
+- Added the following tools:
+  - `Cmd`
+  - `Js`
+  - `Log`
+
+### Changed
+
+- Added a common file (`tests\common.php`) for the autoloading & other common functions needed by the test files.
+
+---
+
 ## Version 1.0.1 - 2022-04-27
 
 ### Added